Phong2004

+++ removeingredient.cpp
@@ -1,5 +1,6 @@
 #include "removeingredient.h"
 #include "ui_removeingredient.h"
+#include <QMessageBox>
 
 RemoveIngredient::RemoveIngredient(QWidget *parent) :
     QDialog(parent),
@@ -11,11 +12,16 @@ RemoveIngredient::RemoveIngredient(QWidget *parent) :
 
 RemoveIngredient::~RemoveIngredient()
 {
+    qDebug() << "~RemoveIngredient()";
     delete ui;
 }
 
 void RemoveIngredient::on_buttonBox_accepted()
 {
+    if (validateUserInput() == false) {
+        return;
+    }
+    // query method to delete ingredient based on input name
     QString ingredientNameToDelete = ui->ingredientNameLineEdit->text();
     QString deleteQuery = "delete from Ingredient where IngredientName = :name";
     QSqlQuery query(database);
@@ -25,6 +31,28 @@ void RemoveIngredient::on_buttonBox_accepted()
     query.finish();
     query.clear();
     qDebug() << "Last error: " << query.lastError().text();
+    QMessageBox::information(this, "Success", "Remove successfully!");
     ui->ingredientNameLineEdit->clear();
 }
 
+bool RemoveIngredient::validateUserInput() {
+    QString ingredientToDelete = ui->ingredientNameLineEdit->text();
+    // check if input is void
+    if (ingredientToDelete.isEmpty()) {
+        QMessageBox::warning(this, "Warning", "You are removing all ingredients!");
+    }
+    // check if no ingredient name match the input
+    // check if no recipe name match the input
+    QSqlDatabase database = QSqlDatabase::database("DB0");
+    QSqlQuery query(database);
+    query.prepare("select IngredientName from Ingredient");
+    query.exec();
+    while(query.next()) {
+        QString existedRecipeName = query.value(0).toString();
+        if (ingredientToDelete == existedRecipeName) {
+            return true;
+        }
+    }
+    QMessageBox::critical(this, "Error", "No ingredients'names matched your input!");
+    return false;
+}
